+++ cpp/cluster-controller/access-control/LocalDomainAccessController.cpp
@@ -159,7 +159,7 @@ bool LocalDomainAccessController::hasRole(const std::string& userId,
                                           const std::string& domain,
                                           Role::Enum role)
 {
-    JOYNR_LOG_DEBUG(logger, "execute: entering hasRole");
+    JOYNR_LOG_TRACE(logger, "execute: entering hasRole");
 
     // See if the user has the given role
     bool hasRole = false;
@@ -186,7 +186,7 @@ void LocalDomainAccessController::getConsumerPermission(
         TrustLevel::Enum trustLevel,
         std::shared_ptr<IGetConsumerPermissionCallback> callback)
 {
-    JOYNR_LOG_DEBUG(logger, "Entering getConsumerPermission with unknown operation");
+    JOYNR_LOG_TRACE(logger, "Entering getConsumerPermission with unknown operation");
 
     // Is the ACL for this domain/interface available?
     std::string compoundKey = createCompoundKey(domain, interfaceName);
@@ -237,7 +237,7 @@ Permission::Enum LocalDomainAccessController::getConsumerPermission(
         const std::string& operation,
         TrustLevel::Enum trustLevel)
 {
-    JOYNR_LOG_DEBUG(logger, "Entering getConsumerPermission with known operation");
+    JOYNR_LOG_TRACE(logger, "Entering getConsumerPermission with known operation");
 
     boost::optional<MasterAccessControlEntry> masterAceOptional =
             localDomainAccessStore->getMasterAccessControlEntry(
@@ -525,7 +525,7 @@ void LocalDomainAccessController::unregisterProvider(const std::string& domain,
         subscriptionIds = aceSubscriptions[compoundKey];
     }
 
-    JOYNR_LOG_DEBUG(logger,
+    JOYNR_LOG_TRACE(logger,
                     "Unsubscribing from ACL broadcasts for domain {}, interface {}",
                     domain,
                     interfaceName);
@@ -691,10 +691,10 @@ void LocalDomainAccessController::initialised(const std::string& domain,
 void LocalDomainAccessController::abortInitialisation(const std::string& domain,
                                                       const std::string& interfaceName)
 {
-    JOYNR_LOG_INFO(logger,
-                   "Removing outstanding ACL requests for domain {}, interface {}",
-                   domain,
-                   interfaceName);
+    JOYNR_LOG_TRACE(logger,
+                    "Removing outstanding ACL requests for domain {}, interface {}",
+                    domain,
+                    interfaceName);
 
     std::string compoundKey = createCompoundKey(domain, interfaceName);
     std::vector<ConsumerPermissionRequest> requests;
@@ -867,7 +867,7 @@ void LocalDomainAccessController::DomainRoleEntryChangedBroadcastListener::onRec
     } else {
         parent.localDomainAccessStore->removeDomainRole(changedDre.getUid(), changedDre.getRole());
     }
-    JOYNR_LOG_DEBUG(parent.logger, "Changed DRE: {}", changedDre.toString());
+    JOYNR_LOG_TRACE(parent.logger, "Changed DRE: {}", changedDre.toString());
 }
 
 void LocalDomainAccessController::DomainRoleEntryChangedBroadcastListener::onError(
@@ -895,14 +895,14 @@ void LocalDomainAccessController::MasterAccessControlEntryChangedBroadcastListen
 {
     if (changeType != ChangeType::REMOVE) {
         parent.localDomainAccessStore->updateMasterAccessControlEntry(changedMasterAce);
-        JOYNR_LOG_DEBUG(parent.logger, "Changed MasterAce: {}", changedMasterAce.toString());
+        JOYNR_LOG_TRACE(parent.logger, "Changed MasterAce: {}", changedMasterAce.toString());
     } else {
         parent.localDomainAccessStore->removeMasterAccessControlEntry(
                 changedMasterAce.getUid(),
                 changedMasterAce.getDomain(),
                 changedMasterAce.getInterfaceName(),
                 changedMasterAce.getOperation());
-        JOYNR_LOG_DEBUG(parent.logger, "Removed MasterAce: {}", changedMasterAce.toString());
+        JOYNR_LOG_TRACE(parent.logger, "Removed MasterAce: {}", changedMasterAce.toString());
     }
 }
 
@@ -938,7 +938,7 @@ void LocalDomainAccessController::MediatorAccessControlEntryChangedBroadcastList
                 changedMediatorAce.getInterfaceName(),
                 changedMediatorAce.getOperation());
     }
-    JOYNR_LOG_DEBUG(parent.logger, "Changed MediatorAce: {}", changedMediatorAce.toString());
+    JOYNR_LOG_TRACE(parent.logger, "Changed MediatorAce: {}", changedMediatorAce.toString());
 }
 
 void LocalDomainAccessController::MediatorAccessControlEntryChangedBroadcastListener::onError(
@@ -973,7 +973,7 @@ void LocalDomainAccessController::OwnerAccessControlEntryChangedBroadcastListene
                 changedOwnerAce.getInterfaceName(),
                 changedOwnerAce.getOperation());
     }
-    JOYNR_LOG_DEBUG(parent.logger, "Changed OwnerAce: {}", changedOwnerAce.toString());
+    JOYNR_LOG_TRACE(parent.logger, "Changed OwnerAce: {}", changedOwnerAce.toString());
 }
 
 void LocalDomainAccessController::OwnerAccessControlEntryChangedBroadcastListener::onError(