+++ cpp/cluster-controller/mqtt/MosquittoSubscriber.cpp
@@ -103,7 +103,7 @@ void MosquittoSubscriber::run()
         }
     }
 
-    JOYNR_LOG_DEBUG(logger, "Try to disconnect Mosquitto Connection");
+    JOYNR_LOG_TRACE(logger, "Try to disconnect Mosquitto Connection");
     int rc = disconnect();
     if (rc == MOSQ_ERR_SUCCESS) {
         JOYNR_LOG_DEBUG(logger, "Mosquitto Connection disconnected");
@@ -261,11 +261,12 @@ void MosquittoSubscriber::on_subscribe(int mid, int qos_count, const int* grante
 
 void MosquittoSubscriber::on_message(const struct mosquitto_message* message)
 {
-    JOYNR_LOG_DEBUG(logger, "Received raw message len: {}", message->payloadlen);
-
     std::string jsonObject(static_cast<char*>(message->payload), message->payloadlen);
 
-    JOYNR_LOG_DEBUG(logger, "Received raw message: {}", jsonObject);
+    JOYNR_LOG_DEBUG(logger,
+                    "Received raw message with len: {}, content: {}",
+                    message->payloadlen,
+                    jsonObject);
 
     if (onTextMessageReceived) {
         onTextMessageReceived(jsonObject);