+++ cpp/common/joynr-messaging/JoynrMessage.cpp
@@ -184,7 +184,7 @@ void JoynrMessage::setHeader(const std::map<std::string, std::string>& newHeader
     while (i != newHeaders.end()) {
         if (!containsHeader(i->first)) {
             header.insert(std::pair<std::string, std::string>(i->first, i->second));
-            JOYNR_LOG_DEBUG(logger, "insert header: {} = {}", i->second, i->first);
+            JOYNR_LOG_TRACE(logger, "insert header: {} = {}", i->second, i->first);
         } else {
             header[i->first] = i->second;
         }