+++ cpp/libjoynr/proxy/Arbitrator.cpp
@@ -22,6 +22,8 @@
 #include <vector>
 #include <chrono>
 
+#include <boost/algorithm/string/join.hpp>
+
 #include "joynr/exceptions/JoynrException.h"
 #include "joynr/exceptions/NoCompatibleProviderFoundException.h"
 #include "joynr/Logger.h"
@@ -54,60 +56,90 @@ Arbitrator::Arbitrator(
           arbitrationError("Arbitration could not be finished in time."),
           arbitrationStrategyFunction(std::move(arbitrationStrategyFunction)),
           participantId(""),
-          arbitrationFinished(false)
+          arbitrationFinished(false),
+          arbitrationRunning(false),
+          keepArbitrationRunning(false),
+          arbitrationThread()
 {
+}
+
+Arbitrator::~Arbitrator()
+{
+    keepArbitrationRunning = false;
+
+    if (arbitrationThread.joinable()) {
+        arbitrationThread.join();
+    }
 }
 
 void Arbitrator::startArbitration(
         std::function<void(const types::DiscoveryEntryWithMetaInfo& discoveryEntry)> onSuccess,
         std::function<void(const exceptions::DiscoveryException& exception)> onError)
 {
+    if (arbitrationRunning) {
+        return;
+    }
+
+    arbitrationRunning = true;
+    keepArbitrationRunning = true;
+
     onSuccessCallback = onSuccess;
     onErrorCallback = onError;
 
-    Semaphore semaphore;
-    arbitrationFinished = false;
+    arbitrationThread = std::thread([this]() {
+        Semaphore semaphore;
+        arbitrationFinished = false;
 
-    // Arbitrate until successful or timed out
-    auto start = std::chrono::system_clock::now();
+        std::string serializedDomainsList = boost::algorithm::join(domains, ", ");
+        JOYNR_LOG_DEBUG(logger,
+                        "DISCOVERY lookup for domain: {}, interface: [{}]",
+                        serializedDomainsList,
+                        interfaceName);
 
-    while (true) {
-        attemptArbitration();
+        // Arbitrate until successful or timed out
+        auto start = std::chrono::system_clock::now();
 
-        if (arbitrationFinished) {
-            return;
+        while (keepArbitrationRunning) {
+            attemptArbitration();
+
+            if (arbitrationFinished) {
+                return;
+            }
+
+            // If there are no suitable providers, retry the arbitration after the retry interval
+            // elapsed
+            auto now = std::chrono::system_clock::now();
+            auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(now - start);
+
+            if (discoveryQos.getDiscoveryTimeoutMs() <= duration.count()) {
+                // discovery timeout reached
+                break;
+            } else if (discoveryQos.getDiscoveryTimeoutMs() - duration.count() <=
+                       discoveryQos.getRetryIntervalMs()) {
+                /*
+                 * no retry possible -> wait until discoveryTimeout is reached and inform caller
+                 * about
+                 * cancelled arbitration
+                 */
+                semaphore.waitFor(std::chrono::milliseconds(discoveryQos.getDiscoveryTimeoutMs() -
+                                                            duration.count()));
+                break;
+            } else {
+                // wait for retry interval and attempt a new arbitration
+                semaphore.waitFor(std::chrono::milliseconds(discoveryQos.getRetryIntervalMs()));
+            }
         }
 
-        // If there are no suitable providers, retry the arbitration after the retry interval
-        // elapsed
-        auto now = std::chrono::system_clock::now();
-        auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(now - start);
-
-        if (discoveryQos.getDiscoveryTimeoutMs() <= duration.count()) {
-            // discovery timeout reached
-            break;
-        } else if (discoveryQos.getDiscoveryTimeoutMs() - duration.count() <=
-                   discoveryQos.getRetryIntervalMs()) {
-            /*
-             * no retry possible -> wait until discoveryTimeout is reached and inform caller about
-             * cancelled arbitration
-             */
-            semaphore.waitFor(std::chrono::milliseconds(discoveryQos.getDiscoveryTimeoutMs() -
-                                                        duration.count()));
-            break;
+        // If this point is reached the arbitration timed out
+        if (!discoveredIncompatibleVersions.empty()) {
+            onErrorCallback(
+                    exceptions::NoCompatibleProviderFoundException(discoveredIncompatibleVersions));
         } else {
-            // wait for retry interval and attempt a new arbitration
-            semaphore.waitFor(std::chrono::milliseconds(discoveryQos.getRetryIntervalMs()));
+            onErrorCallback(arbitrationError);
         }
-    }
 
-    // If this point is reached the arbitration timed out
-    if (!discoveredIncompatibleVersions.empty()) {
-        onErrorCallback(
-                exceptions::NoCompatibleProviderFoundException(discoveredIncompatibleVersions));
-    } else {
-        onErrorCallback(arbitrationError);
-    }
+        arbitrationRunning = false;
+    });
 }
 
 void Arbitrator::attemptArbitration()